extract out helper func, rework uploaded media filter
